some minor updates.

+++ types.h
@@ -7,18 +7,18 @@
 
 #include <cstring>
 
-typedef unsigned char   uint_8;
-typedef unsigned short  uint_16;
-typedef unsigned int    uint_32;
-typedef unsigned long   uint_64;
+typedef unsigned char           uint_8;
+typedef unsigned short          uint_16;
+typedef unsigned int            uint_32;
+typedef unsigned long long      uint_64;
 
-typedef char            int_8;
-typedef short           int_16;
-typedef int             int_32;
-typedef long            int_64;
+typedef char                    int_8;
+typedef short                   int_16;
+typedef int                     int_32;
+typedef long                    int_64;
 
-typedef float           flt_32;
-typedef double          flt_64;
+typedef float                   flt_32;
+typedef double                  flt_64;
 
 typedef unsigned char*  pointer;
 typedef unsigned long   pointer_literal_type;
